Add better semantics for `if` expressions:
If expressions are now checked to make sure they have boolean-compatible
conditions, that they actually have a body, etc. I also fixed a lot parser
errors relating to whitespace, since if expressions introduce nested blocks
which need to handle whitespace/newlines differently then it used to.

One issue with this that cannot easily be fixed is the fact that an if expr
which has no else body (like the ones currently implemented) shouldn't be
able to be assigned to a variable, since an `if` expr that is never executed
will never have a type. We can fix this by introducing an `Option` type similar
to Rust, or a None/null type union like in Python. I like the rust option more
because it allows you to do stuff like `x.value_or(...)` and similar. This
would be a lot of work though, too much for this commit.