Add union types:
There is no way to explicitly define these union types, but they are currently
being generated in if expressions to account for the fact that they might have
an unknown value. This will become a lot more useful when if exprs can have
else and elif bodies, but for now there is a union type which can be used for
comparisons and so on. They only implement `__eq__`, but can be updated to
support `__lt__` and `__gt__` for things like subsets.