Add better user/repository permissions:
Previously Cicada would only grant a user "owner" permissions if the sender
matched the username of the repository owner (and no other permissions would
be granted). Now any user who is allowed to push code will be granted write
access (in Cicada), and anyone who can open an issue will be given read
access. This is far from ideal, but is a starting point in terms of supporting
organization repos, since the owner of the repository is the organization,
not the user who triggered the webhook.