Simplify AST parser state machine:
Previously the parser used an iterable stream of tokens instead of a list of
tokens. The reason behind this is that I wanted to construct the AST tree as I
was parsing the file, that way we wouldn't tokenize any extra data in case of
an AST generation error. This provided little advantage and created lots of
extra work for the parser since it had to manage the tokens it had alread seen
in case we needed to rewind the parser.