Rewrite pre-tokenizer:
Similar to the AST generator, the chunk grouper (the stage before the
tokenizer) used to lazily group the chunks into tokens. This made it hard to
group certain tokens because we would occasionally need to rewind, meaning
the logic for grouping the chunks was hard to comprehend and hard to maintain.
Now the chunk grouping logic uses a stateful parser like the AST generator.