Add installation abstraction layer:
Because Cicada can be installed from lots of different providers it can be hard
to figure out what the "real" instance of Cicada actually is, and harder yet,
which one to bill for. Although this isn't fully fleshed out, this provides a
starting point for future commits. Essentially an installation is now created
whenever a webhook is recieved saying a new installation (user or organization
scoped) has been added to Cicada. We assume that the user who triggered the
webhook is the admin. Eventually admins will be able to add/remove users from
instances, which should provide a better experience when it comes to using
Cicada.