Add page for managing installations:
Basically just add a page for viewing an existing installation. Currently the
only options are to view the installation on the providers site (such as
GitHub), thought this page will be updated in the future to allow the creation
of users and so on.

Also, because multiple installations can be installed to the same location I
had to change the unique constraints on the installations table. An
installation is now considered unique if the name, provider, and provider info
(url and ID) are distinct.