Speed up page loading:
* Deduplicate CSS and JS
* Remove unused CSS classes
* Add gzip middleware to FastAPI

To speed things up even further we should minify the HTML/CSS/JS/SVG files,
and also move the static content to NGINX (or a separate CDN all together).

There should also be a favicon file since the browser makes a request for one
every page, which is wasteful because it is always a 404, meaning it has to try
again every page load. I have yet to make a logo for Cicada, so once I do that
I'll add it here.