Add support for floating point numbers:
Numeric expressions/values can now support both int and float types. There are
subtle issues with using `Decimal` to represent these types, primarily in how
binary integer operations (such as and/or/xor) work. Because these are normally
done on integers, but these numeric values can work on floats as well, I
decided it would be best to convert the numeric types to ints, do the binary
operation, then convert it back to a numeric (`Decimal`). This should be
documented once I add docs for the operators.