Fix data still being sent after websocket was closed:
This was a weird bug which was in part due to the uvicorn library wrapping the
websocket provided by fastapi/starlette in addition to the websocket using by
`create_task` throwing an exception that was not handled properly.

This is a quick fix, but I will need to wrap this in some sort of wrapper
because using websockets is very verbose currently as I have to handle auth,
check the state of the socket, and so on.