Add view for user session session info:
The code for querying sessions was starting to get a bit out of hand. Before it
relied on lots of helper functions and string concatenation, now the queries
are expanded out to be more readable. Although this more code, I find that it
is much more straightforward as to how the data is queried.

I should really look into a SQLite query builder at some point, but for now
this will suffice.