Add notifications for unhandled exceptions:
Notifications are sent via ntfy.sh, which is a really good service for sending
and recieving notifications without an account. This is optional, and only
enabled if you set the `NTFY_NOTIFICATION_URL` env var. If the notification
cannot be sent a critical message will be printed saying so. Regardless of
whether the message was sent or not, the full traceback is printed to the
screen.