Make permission levels more explicit:
Previously the permission levels required for doing certain actions was never
explicit, and it became hard to ensure the proper permissions where being
checked before calling an application service, endpoint, etc. Now the
permissions are baked in as part of the querying process, meaning that the
required permissions are much more explicit.