Add mutable variables and assignment operator:
By using `mut` after `let` you can defined a mutable variable. A new `=`
operator/expression has been added as well to enable assigning to a variable.
Because this `=` operator is an expression it can be used in conjunction with
other expressions. Because `=` might be confused with `==` or `is`, there
should (in the future) be a warning for potentially ambiguous usage of `=`.

In order to make the assignment expression to work properly I had to change
how functions are parsed: Essentially only the `shell` function and shell
aliases (like `cd`) are allowed now. This isn't an issue since user defined
functions don't exist yet, just something to keep in mind. I might change my
mind on this, but in my opinion (as of now), functions should be used with
parenthesis whenever possible, and only "shell" related commands should be used
with the shell format. Since only the shell format is available right now
there is no changes that need to be made, just something to keep in mind.