Use `Self` type in more places:
This is a Python 3.11 feature, but can be used by importing from
`typing_extensions`. It isn't a silver bullet, but it does greatly clean up
the code, and removes the need for `from __future__ import annotations`. There
was one place where I had to keep the future import and that was in the node
file due to the `NodeVisitor` and the `Node` classes depending on one another.
I could've used quoted types instead, but opted to just leave it the way it
is.