Add dataclass builder for tests:
I was tired of building test domain objects from scratch, so I wrote a little
builder function to inject default data into dataclasses. This will need to be
built upon, but right now it solves most of the dataclass usage in this
project. I will need to go back through and replace some of the functions
(namely `create_dummy_session()` and `User()` objects) with the new builder.