Add new "remote podman" executor:
This remote podman executor is a stepping stone towards self-hosted runners.
Previously workflows had to be executed inside of a container with that had
Cicada installed in it, meaning you could not easily use custom containers.
Now you are (in theory) able to use any container you want without having
to set it up for use with Cicada.

There is still some work to do here, namely allowing users to define the
container that they want to use in the CI file itself (currently the remote
executor is hard coded to the `alpine` image).

This executor is better in many ways:

* It only has to download the project once, and uses a mount inside the
  container.
* In theory, any container can be used (will be finalized in the future)
* Less moving parts
* Doesn't require pre-building any images

To make all of this work, a few (small) things had to change in Cicada:

* Each repository was temporarily cloned during the gather phase, and cloned
  again in the runner phase (inside the container). This has been eliminated
  (in the remote runner) since the cloned repository is now available in
  both the gather and runner phases.

* The SQLite database is potentially not thread-safe. This was needed to
  allow for the container running in the async thread to write stdout data back
  to the terminal object, which is stored in SQLite. There should only be one
  service writing to this object, so it should be fine. This does mean though
  that future thread related issues might not be caught now.