Add public repository support:
Previously all repositories where treated as private, so whenever you tried to
see the output for a session that was a public repo, you would be unable to.
Now whenever an event is pushed, Cicada will check if the repository is public
or not, and store that along with the other repository information. Now any
user can view the run data for sessions triggered on public repos. There should
probably be a way to disable this behaviour for people who dont want/like it,
but I will cross that bridge when I get there.

Currently this public repository detection is only supported on GitHub. In
general, Gitlab support is lacking compared do GitHub, simply because you need
to manually setup webhooks for every repository/project you want to use Cicada
for, which makes testing and integrating annoying. Adding better Gitlab support
is planned, but not the highest priority atm.