Upgrade to Python 3.11:
Since the server is running on Python 3.11 it only makes sense that my local
machine is on 3.11 as well. There is now an issue/warning with the `passlib`
module saying that the `crypt` module is deprecated. I find it sad that this
module is going away, though I imagine people rarely interact with user data
through Python anyways.