Add new `run_on` statement:
This `run_on` statement allows for telling Cicada which OCI compatible image to
use for running the workflow. Currently only OCI images are allowed, which is a
good start, but eventually this will be expanded to allow for self hosted
runners and other runner mechanisms.

The `run_on` statement has similar semantics to the `on` statement. See the
docs for a more in-depth explaination.