Fix remote executor picking unrunnable `.ci` files:
This repository includes `.ci` files for testing (which are empty), and the
code to find which `.ci` files to run was including all files, not runnable
ones. I already created a function to do just that, which is used in the
gather phase. This list is not passed down from the gather phase, though it
probably should, since we are just recalculating it here again. Since Cicada
doesnt support multiple workflows for a single session, only one workflow file
can be ran at a time. This is starting to become an issue so I will need to
address it soon.