Begin to add workflow logic:
Currently "workflows" and "sessions" are synonymous, but they shouldn't be: A
workflow is a single output of a workflow file run, and a session is an object
that encapsulates that. Currently there is a concept of "rerunning" a session,
but even this is a hack because we create a new session with the same id, but
increase a run counter. This new system will create a single session and make
workflows that live under a single session, as opposed to a bunch of similar
sessions.

This is not at all complete, but I dont want to leave this hanging, so I am
merging it in now (and it doesnt do anything as of yet, so no harm in letting
it do its thing). I'm doing this to allow for multiple workflows in a single
session, and to do that I have to keep track of the filename that runs the
session, while also keeping track of the runs for the current session. This
adds a lot of complexity, something that the old system isn't capable of
expressing.

The next step is going to be adding more workflow repo methods, since those are
what I usually want when I am dealing with workflows (such as the workflow
output page). After that I plan on re-attaching the terminal sessions to the
workflows, since those will be a one-to-one mapping, then start to rehash the
old sessions so that they fit the new system.