Fix new remote runner not showing command output in real time:
Because of the way I am calling podman now (via `podman exec`) I was running
into issues where podman couldn't tell I was using a TTY, and because of this,
was not buffering the output correctly. The man pages mention something about
the `-t` flag potentially "hanging pipes", though I don't know why that should
change if you're using a TTY. Regardless, the `-t` flag is needed to allow for
color and other escape sequences, so I had to do a lot of hacks to make the raw
TTY output show up.

A side effect of this is now there are a lot more carriage returns in stdout,
which I find weird. Regardless, ansi_up does not currently support carriage
returns and uses a newline instead. This was ok before, but now that there is
much more carriage returns, I had to replace consecutive CRs with a single CR.
There is still an issue where an extra newline is added where it shouldn't be,
but I am happy with what I have so far. I might look into fixing ansi_up if it
isn't too hard, but now that everything is working I probably won't touch it
right away.