Set TTY column size to 120:
Previously the TTY sessions would use the column size of the TTY that was
running the server code, which isn't stable, and in my case, was set to 212,
way too long to be useful for CI output. I have hard coded this to 120 which
is long, but not too long to cause display issues like before.

I could (and probably should) hard code the column size on xterm.js, though
I believe it is fine as long as there are >120 columns in the browser. This
might be an issue if you have Cicada in a small browser window or on mobile,
so I will have to look into this a bit more.