Fix session status not getting updated properly:
Previously we would wait for a small duration then check the session status
to see if the status was updated. Chances are that the session would not be
updated by then, meaning we would just send a pending state back to the client.
Now we recheck multiple times over a 5 second interval to ensure the status
is sent back properly. We timeout after 5 seconds in case of a bug which causes
the session to never finish.