Add tool to convert GitHub Actions workflows to Cicada workflows:
This is a very experimental tool/feature which should help lower the barrier of
entry for using/adopting Cicada. Due to the highly complex nature of CI/CD
workflows, most workflows will need some tweaking to make them actually work,
but the goal of this tool (for now) is to give users an example of what their
workflows could look like with Cicada without forcing to do a bunch of work up
front.