Add support for `.cicadaignore` file:
Cicada will now ignore any folder which has a `.cicadaignore` file in it. This
is useful for ignoring test data files, otherwise Cicada would treat them as
real workflow files and run them, which is not ideal. A better, more long term
approach would be to make the `.cicadaignore` file act like a `.gitignore` file
in which you specify globs for the files/folders you want to ignore.