Remove `convertEol` flag:
This flag was adding newlines where there shouldn't be which caused strange
output in weird places. When refreshing these artifacts go away, which leads
me to believe that the issue is due to partial streams starting/ending with
CR/LF characters causing extra newlines to be added. When refreshing, all the
stdout is returned at once, and there are no such artifacts.