Add ability to parse C style functions:
The shell style is good for command like operations, but for more data
orientated work it is better to have C style function calls. This makes it
easier to figure out which args are which, and to allow for multiple function
calls on a single line (currently shell style function calls need to be on
their own lines). Currently these C style functions cannot be used with the
shell functions like `git` due to the fact that these shell functions have
higher precedence, thought this might change in the future. Some work will
need to be done to the `FunctionExpression` class because the `name` field
cannot support member functions such as `"abc".to_upper()` or similar, since
the function is being called on a string, not at the global scope.