Replace old landing page with new one:
This has been in the works for a while, but I just put on the final touches
today. This new site gives a much better description of the product itself,
as well as includes pricing for the different tiers (free, cloud, and custom).

Here are some of the highlights:

* Add screenshot
* Add logo and favicon
* Add lots of neon
* Add workflow comparison to GitHub Actions and Gitlab CI
* Add company info and copyright at the bottom

One of the things I removed was the waiting list feature: Since Cicada is
already live it doesn't make sense to have a waiting list anymore. I might add
it back in for people who are agnostic and want to stay updated.

Some of the things I still have to work on are the privacy policy and the terms
and conditions. Currently the links are blank, which is not ideal, but creating
these legal documents is something you need to take seriously, so until I get
them to a place I'm happy with I'll keep the links blank.