Store user's email when logging in via GitHub SSO:
In order to bill users for using Cicada we have to keep track of their email
address. Cicada will use whatever email is set as your primary email on GitHub.
Currently there is no way to change this, but it should be easy enough to add
a "change billing email" button later on.