Replace `Final` with `ClassVar`:
I originally assumed `Final` meant that the value could not change and was
immutable (at least when statically analyzing), but that is not the case at
all. Values can still be modified, but they cannot be reassigned. `ClassVar`
acts similarly, though since I am using this value in a class, it makes more
sense to use `ClassVar`.