Fix favicons not working:
Originally I was using `/static` as the route for the static files, but this
was problematic because I couldn't use `/`, which is needed for the favicon
files. The issue was that I was mounting the `/static` path near the top of the
file, meaning it would take precedence over all the `/*` paths. By moving it
to the bottom I can use static files on in the `frontend` folder while still
using `/` for API stuff.