Fix static file router causing tests to fail:
Now that the static files are mounted at `/`, all 404 paths are checked to see
if they are files. This means that the POST request is expected to be a GET,
since the path didn't exist. This isn't ideal, since any invalid path, POST or
not, should be a 404.

The real solution would be moving all the static content to NGINX, though the
FastAPI client is nice for local dev testing when you don't want/need NGINX.