Add placeholder env vars to mimic GitHub default env vars:
Lots of scripts/workflows depend on GitHub env vars such as `GITHUB_SHA`, so
until Cicada can properly detect and convert these env vars in GitHub Action
workflows it is best to inject these values via env vars. It is quite trivial,
and improves compatibility with other platforms, so a win win. Also, the `CI`
env var is set to `true` (which GitHub and Gitlab both do) as it is a quick
way to let scripts know if they are running on a CI server or not.