Add container improvements/fixes:
Images that don't include a version will default to the `latest` version. Now
when an image is not found a proper error message is displayed to the terminal,
and the session stops. This could be done in a more elegant way, but for now
this should at least let the user know that something went wrong.