Replace default `alpine` image with custom image:
The built-in `alpine` image was lacking, and didn't have a lot of the features
that you'd typically expect. There are a lot of packages pre-installed now,
including `bash`, `git`, `sqlite`, `gcc` (and related tools), and more. The
image is about 270MB, which is a bit bigger than I would've liked, but it will
have to do for now.

One of the issues I just realized is that even though I installed `bash`, `sh`
is still being used as the default shell. The hard part is that I can't know
if `bash` is installed for an image just by looking at it. For now I'll just
keep using `sh`, though I'll need to find a way to use `bash` if it is
available.