Restructure and rename files:
The codebase is overdue for some routine cleanups. The folder structures have
gotten a bit out of hand, the file names don't always represent what they
contain, and the certain bits of code are in places that they shouldn't.

This is the first of many commits aimed at cleaning up and restructuring the
project in a more comprehensive manner.

Overview of this commit:

* Remove unused issue CI file
* Move `cicada.api.domain` to `cicada.domain`
* Move `cicada.api.repo` to `cicada.domain.repo`
* Move `PasswordHash` class to `cicada.domain.user`
* Move `cicada.api.application` to `cicada.api`
* Relocate `run_check` file and `EnvironmentVariable` class
* Relocate `cicada.api.common.datetime` to `cicada.domain.domain`
* Relocate `cicada.api.common.http` to `cicada.api.infra.common`
* Relocate `cicada.api.common.json` to `cicada.common.json`
* Remove `GeneratorWrapper`