Add self hosted runners:
In short, self hosted runners are runners that you can run on prem, but talk
to the live cicada.sh site. It bridges the gap between people who don't want to
self host Cicada but still need to run automation on prem.

There are lots of things in this squashed commit, which I will explain here:

I needed to change lots of things in Cicada to get this working in the first
place. This includes:

* Passing the AST tree for each workflow to the execute stage so that a self
  hosted executor can be used for the new self hosted runners.

* Inject a DI container into workflow runners. Previously the workflow runners
  would just do their thing and stream the results back, but now that there
  is a runner involved there is more information that needs to be sent back
  to Cicada which previously wasn't needed, like notifying Cicada when the
  workflow has been picked up by the runner.

Once I got these refactors out of the way I was able to actually get to work
on the runners themselves. The general gist of how the runners work is as
follows:

* User causes a trigger like normal, but uses the `self-hosted` runner type.

* During the execute phase, if any workflow is has this new self-hosted runner
  type, the new self hosted execution environment will be used to run the
  workflow.

* This new execution environment will queue the session and AST tree so that
  it can be picked up by the runner (more on this later). Once the session has
  been queued we wait until the session status indicates the session has been
  finished, then exit.

* There is a new "runner endpoint" where runners can connect to a websocket
  endpoint on Cicada. Once a runner is authenticated to our endpoint, the
  server will check the session queue for any new sessions for the connected
  runner. If a new session is available, all the workflows in that session
  will be sent to the runner. The server will wait for each of the workflows
  to finish, making sure to capture and relay any stdout information sent via
  the runner.

* On the runner side, it will connect to the runner endpoint using it's runner
  ID and shared secret. Once authenticated, the runner will sit idle and wait
  for new workflows to run. When a runner recieves a workflow it will run it,
  stream all the stdout data back to the server, and once the workflow has
  finished, it will notify the server that the workflow is finished.

To bring it all together, you need to install the runner on a client machine.
To do this, I had to split up and package parts of Cicada so that they could
be installed on a local user's machine. This was harder then I expected, and
has some caveats:

* The `cicada` package is already taken on PyPi, and I was having issues with
  changing the name of the folder when building, the `cicada` package will have
  to be installed manually.

* The `cicada-runner` package is a small package that only contains the code
  for running the server, and only depends on `cicada` (and a few 3rd party
  ones as well). While this makes packaging simpler (in some aspects), it
  also means the end user has to install 2 packages at once.

* Because Cicada is structured as a mono-repo (for now), it is very hard to
  build the `cicada` and `cicada-runner` packages. The process isn't very
  intuitive, and will be improved upon in the future.

* The `cicada` and `cicada-runner` packages aren't published, and aren't able
  to be downloaded. The code for the runner lives in this repo, but the actual
  packages themselves are not available.

Lastly, and it goes without saying: these new runners are highly experimental
and subject to change. As it stands right now there is no user interface for
adding/removing/modifying runners, so you need to manually import them and
this process has yet to be documented. Only a few people will even be using
these runners to start off with, so the least amount of energy is being put
into making a fully fledged solution, and instead a small MVP is being built
instead.

This commit message is getting a bit long, but a short TODO list to end things:

* Add documentation (setting up runner, adding runner, etc).

* Improve/robustify system for sending/recieving workflows to runner.

* Improve test coverage for the new code that was added.

Note: This is about 50 commits worth of code that really should've been worked
on in small sections, but instead was left on a branch for too long. Next
feature will be much smaller!