Add Docker image for running self-hosted runners:
This provides an easy way for people to start using Cicada self hosted runners
on their own hardware. Due to how Docker works, they won't get access to the
host machine, but they will get access to the host network, which might be
enough.

This Dockerfile also acts as self-documenting instructions for how to setup and
run a self-hosted runner.