Fix last commit:
It would appear that Mypy is using a cache even when a class that is being
imported has been renamed. Perhaps this is because I am checking the `cicada`
and `test` folders in separate commands, though I doubt that is the reason.

Adding a `clean` recipe to the Makefile to clear out caches. Ideally this
wouldn't happen in the first place, but at the very least I can quickly clear
out my cache before testing and pushing.