Fix display name being used instead of username in `git.push` triggers:
This display name is not the display name you see on GitHub, but the display
name you have set in your local git config. This could cause issues later on,
for example, if a commit is pushed that has a username/email that cannot be
attached to a GitHub user, the `username` field might not exist.