Add `run_on_self_hosted` flag for Session and Workflow objects:
This allows for better debug-ability, and in the future, a way to exclude
self-hosted sessions from billing (since users using their own hardware should
not be charged for it).

Some technical things that I need/should address at some point: It is a real
pain to have to drop and recreate views each time I want to change them. Using
a proper ORM and migration system will probably fix that. Also, the
`run_workflow` functions should take an object parameter instead of a bunch of
assorted parameters. These params change very often, and the tests depend on
these, so having to update the tests (which only mock the inputs and don't
actually use them) is really annoying and time-consuming.