Bump packages, add `typos` spell checker:
I recently was introduced to `typos`, which is a really great spell checking
tool (written in Rust, of course). I probably should've made this as a separate
commit but oh well.

I don't like having to make a one-off `.typos.toml` file for this, but I
intend to open a PR to fix it soon.