Fix soft-crash when `podman` is not installed:
When using the `remote-podman` executor (currently the only executor), Cicada
will throw an exception if `podman` is not installed. This causes the session
to fail, and it gets in a hard-to-recover state. This fixes that.

The next step is to add a normal `remote-docker` executor, though this will
require running the server as `root`, or require that the current user is in
the `docker` group.