Verify env vars on startup:
This will help identify misconfigurations earlier. Currently env vars are lazy
loaded when they are needed which means you won't know if they are valid or not
until they are used. To help mitigate env vars are all loaded once when the app
first starts to make sure the env vars are valid, and then they are lazy loaded
as normal afterwards.