Fix websocket not stopping session if it's in an invalid state:
Sometimes when an error occurs a session gets into a broken state that it can't
recover from. Because the session is still "live" the websocket remains open,
but the server is not there to respond. On the frontend the client sends the
STOP command via the websocket (if available), but since the server is not
responding, the command is ignored. This commit adds a timeout to send a force
stop command if the websocket is non-responsive. More effort should be made to
make the server itself more robust, but this is a good fix for the time being.