Fix last commit:
Turns out Poetry does not play well with setuptools and setup.py files in
general. In general I don't like poetry because it forces you to use a specific
structure and format which simply does not work for this project.

Also add `typos` to `all` target in Makefile because I forgot to add it.