Fix coverage issues:
Turns out my e2e tests where hitting the email service on my local machine but
not on CI. The email service was slightly increasing coverage on my machine,
but dipped just low enough on the CI server. Fixed by adding a mock to stop
connecting to the email server during tests, and ignore coverage on the
function itself.