Cleanup member expr assignments:
The code I wrote last night wasn't as robust/clean as I would've liked it, so I
went back and refactored it so that it to cover more edge cases. There is a lot
more refactoring to go, I feel like it would be best break each visitor
function into it's own function so that there is less code to look at for any
given visitor.