Add `hashOf` builtin function:
This function will hash one or more files (or file globs) and return a single
hash for all the files. More on the exact semantics can be found in the docs.
The exact hash function used is sha256, though this might change over time. The
hash is made by expanding all the globs, sorting the list of files,
concatenating all the files together, then hashing them. This hashing method
should be good enough, though there could be some issues I'm not seeing right
away (cryptographically speaking).

In working on this feature, some issues/things to consider have arisen:

Currently there are 3 execution environtments:

* Local: Running a `.ci` file locally
* Remote Podman: Inject commands into a running Podman container
* Self hosted: Send commands to a self-hosted Cicada runner

The issue with this is that, for the one `hashOf` function, I had to create
3 different implementations/wrappers, which will not scale in the future.

The reason I can't use the same implementation is that Python isn't guarenteed
to be installed in every OCI image, which isn't a great excuse because the
workaround version I'm using requires `bash` and `sha256sum` to be installed,
which might not be the case. If I was somehow able to mount the underlying
filesystem on the host I could actually use the same executor as the local
CI runner, so that is good to know.

For the self-hosted runner I need to redirect the stdout coming from the print
statements into the terminal output, which isn't too hard, but still annoying
as it requires rebuilding the runner each time.

In addition to the actual running of the `hashOf` functions, the semantic
analyzer needs a bit of an overhaul: It is getting tedious to add things
quickly, mainly because the amount of code you have to sift through to figure
out where to add what you need. I need take a day and split apart all the
functions into smaller ones, and hide the implementation details in other
modules so that the main semantic visitor is clean.