Redirect user to login page if JWT token expires:
Due to the short lifetime of the JWT tokens, the user will often land on a page
that requires authentication, have to log out, log back in, and find the page
that they where on. Now whenever their token expires, they automatically get
redirected to the login page, and once they log back in, they will be sent back
to the page they where originally on.

In addition, the "View more details" URLs will automatically sign the user in
via GitHub (via the SSO url), meaning the user won't need to click "login with
GitHub" when clicking on the URL from GitHub. This might cause issues later
with public workflows, since anyone clicking on that link will be prompted to
login with SSO.