Add caching support:
Users can now use the `cache` command to cache files and folders using a cache
key. This is highly experimental, subject to change, and currently has a lot of
restrictions in place for security and cost reasons.

The basic technical implementation is as follows:

* Cache statement is hit
* Cache key is calculated:
  * If cache key exists, files are loaded and the workflow continues.
  * Otherwise, when/if the workflow succeeds, the specified files are uploaded.
* Cached files are stored in a compressed tarfile, then uploaded to S3

There are a lot more steps than that, but that is the general gist. Look at the
caching docs for more info on how caching works, and what limitations are in
place.