Add placeholder `secret` global variable:
This is the part one in supporting global variables in Cicada. I'm trying to
make small commits now instead of long drawn out feature branches, so there
isn't a lot in this commit. The exact semantics of the `secret` global variable
is not fully ironed out, but that will come soon enough.