Add UI to add secrets:
This doesn't actually save the secrets yet, just placeholder UI until I add the
endpoints and backend for it. I went ahead and updated VanJS as well since it
was a bit out of date. I also fixed a login bug that was preventing you from
logging in with a username/password instead of using SSO.