Fully implement secrets:
This is another one of my mega commits, so there is a lot to unpack here.

This commit adds the following:

* Full support for installation secrets (create/update/delete)
* Full support for repository secrets (also create/update/delete)
* Ability to use secrets in workflows

A lot of magic happened behind the scenes to make this work, but this is a
big step forwards for Cicada.

To acheive this, Cicada stores secrets in Hashicorp Vault. This helps mitigate
the potential security risks associated with managing secrets yourself by using
a widely used and audited program. With that in mind though, here are some of
the technical security measures in place to ensure that secrets in Cicada
remain secure:

* Vault is running on a separate server
* Vault is communicated with via mTLS, meaning only Cicada can talk to it
* Plain text secrets are sent to Vault to be encrypted, then encrypted using
  AES256-GCM96, then sent back to Cicada to be stored until decryption
* The underlying encryption keys auto rotate after 30 days, though Cicada can
  still decrypt secrets that have not been rotated
* Each repository and each installation gets it's own encryption key, ensuring
  that an attacker cannot decrypt every secret if they happen to get the key
  for one secret.

There is still much to be improved, but this is a good starting point going
forwards. While the current implementation should be sufficiently secure, more
security hardening and best practices will be applied on a rolling basis as
they are discovered. This secret implementation has yet to be audited, though
it has been designed with security in mind: Currently, only repository and
installation owners can list, create, or update secrets, and secrets can only
be viewed when they are passed to a workflow.

Lastly, the UI for managing the secrets will probably need to be updated in
the future. While it works, it does not match the env var UI, so one or the
other will need to be updated. In addition, the repository screen should show
the available installation secrets (similar to GitHub) so that you are aware
of what installation-wide secrets already exist.