Add ability to parse basic GitHub Action expressions:
This implementation is very basic and is only tested in a few scenarios, but it
will be improved in the future. Right now the expression parser only works for
command arguments, and will only really work with certain expressions. For
example, and/or exprs will not work, certain globals like `secrets` should be
renamed to `secret`, and so on.