Add optional argument and return types to functions:
This commit enhances functions to allow for optional trailing return types, and
optional type annotations for function arguments, similar to Python. This is
very exciting and something that has been over the horizon in Cicada for a
while now, but is finally getting implemented. Adding user defined functions
allows for more high-level features in the future such as modules.