Replace `RecordField` with `dict`:
This was a premature refactor that made dealing with records much harder. The
idea was that having a `list` of record fields would make it clear that a
record has an ordered list of fields, but that really doesn't matter at this
stage since users can't defined their own records. Even when I do add that
capability, `dict`s in Python are ordered, so there is little reason to use a
`list` anyways.